--- src/containers/cabinet/index.js
@@ -1,3 +1,0 @@
-import Cabinet from './cabinet';
-
-export default Cabinet;
