+++ src/containers/header-profile/index.js
@@ -1,0 +1,3 @@
+import HeaderProfile from './header-profile';
+
+export default HeaderProfile;
